% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_rules.R
\name{has_failures}
\alias{has_failures}
\title{Le rapport contient-il au moins un echec ?}
\usage{
has_failures(report)
}
\arguments{
\item{report}{Un \code{me_check_report}.}
}
\value{
\code{TRUE}/\code{FALSE}.
}
\description{
Le rapport contient-il au moins un echec ?
}
