% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_rules.R
\name{me_check_report}
\alias{me_check_report}
\title{Rapport de controle qualite}
\usage{
me_check_report(
  project_name = character(0),
  results = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names))
    {
         list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })()
)
}
\arguments{
\item{project_name}{Nom du projet controle.}

\item{results}{Un \code{data.frame} avec colonnes \code{rule}, \code{severity},
\code{status} (\code{"OK"} ou \code{"ECHEC"}), \code{message}.}
}
\description{
Rapport de controle qualite
}
