% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_default_rules.R
\name{run_checks}
\alias{run_checks}
\title{Executer un ensemble de regles sur un projet}
\usage{
run_checks(project, rules = default_rules())
}
\arguments{
\item{project}{Un \code{mecore::me_project}.}

\item{rules}{Liste de \code{me_check_rule} (par defaut : \code{default_rules()}).}
}
\value{
Un \code{me_check_report}.
}
\description{
Executer un ensemble de regles sur un projet
}
